refactor: move FTA helpers to subapp
